Add x/y offsets
Because some purchased forms may be laid out slightly differently or
perhaps some printers may print slightly differntly, we add x_offset and
y_offset attributes so that the start position of the placement of text
can be adjusted by the user. Note that, to be consistent with prawn, a
positive y value is higher on the page.